requirements.txt
colorama==0.4.6
iniconfig==2.0.0
packaging==23.1
pluggy==1.3.0
pycodestyle==2.11.0
pytest==7.4.2
PyYAML==6.0.1
